refactory flow struct
