+++ src/model/components/gxl_env/meta.rs
@@ -1,0 +1,9 @@
+use crate::{annotation::EnvAnnotation, meta::GxlType};
+
+#[derive(Clone, Getters, Default)]
+pub struct EnvMeta {
+    class: GxlType,
+    name: String,
+    mix: Vec<String>,
+    annotations: Vec<EnvAnnotation>,
+}
