+++ src/model/components/gxl_flow/meta.rs
@@ -1,0 +1,10 @@
+use crate::{annotation::FlowAnnotation, meta::GxlType};
+
+#[derive(Clone, Getters, Default)]
+pub struct FlowMeta {
+    class: GxlType,
+    name: String,
+    annotations: Vec<FlowAnnotation>,
+    preorder: Vec<String>,
+    postorder: Vec<String>,
+}
