+++ src/model/components/gxl_intercept.rs
@@ -1,5 +1,3 @@
-use crate::ability::prelude::ComponentMeta;
-
 use super::gxl_spc::GxlSpace;
 use super::{prelude::*, GxlFlow};
 
@@ -58,73 +56,5 @@ impl PropsTrait for RgIntercept {
 impl AppendAble<RgProp> for RgIntercept {
     fn append(&mut self, prop: RgProp) {
         self.props.push(prop);
-    }
-}
-
-#[derive(Clone, Getters, Debug)]
-pub struct FlowRunner {
-    m_name: String,
-    flow: GxlFlow,
-    before: RgIntercept,
-    after: RgIntercept,
-}
-impl FlowRunner {
-    pub(crate) fn new(
-        m_name: String,
-        props: Vec<RgProp>,
-        flow: GxlFlow,
-        befores: Vec<GxlFlow>,
-        afters: Vec<GxlFlow>,
-    ) -> Self {
-        Self {
-            before: RgIntercept::new(m_name.clone(), props, befores),
-            flow,
-            after: RgIntercept::new(m_name.clone(), Vec::new(), afters),
-            m_name,
-        }
-    }
-}
-
-impl DependTrait<&GxlSpace> for FlowRunner {
-    fn assemble(self, mod_name: &str, src: &GxlSpace) -> AResult<Self> {
-        Ok(Self {
-            m_name: self.m_name,
-            before: self.before.assemble(mod_name, src)?,
-            flow: self.flow.assemble(mod_name, src)?,
-            after: self.after.assemble(mod_name, src)?,
-        })
-    }
-}
-
-#[async_trait]
-impl AsyncRunnableTrait for FlowRunner {
-    async fn async_exec(&self, mut ctx: ExecContext, dict: VarSpace) -> VTResult {
-        //let orgion = dict.clone();
-        let mut job = Job::from("scope_flow");
-        ctx.append(self.m_name.as_str());
-        // 使用链式调用和模式匹配
-        let dict = {
-            let (d, t) = self.before().async_exec(ctx.clone(), dict).await?;
-            job.append(t);
-            d
-        };
-
-        let dict = {
-            let (d, t) = self.flow().async_exec(ctx.clone(), dict).await?;
-            job.append(t);
-            d
-        };
-
-        let dict = {
-            let (d, t) = self.after().async_exec(ctx.clone(), dict).await?;
-            job.append(t);
-            d
-        };
-        Ok((dict, ExecOut::Job(job)))
-    }
-}
-impl ComponentMeta for FlowRunner {
-    fn com_meta(&self) -> GxlMeta {
-        self.flow().meta().clone()
     }
 }
