+++ src/model/components/mod.rs
@@ -13,8 +13,8 @@ pub mod gxl_var;
 pub mod prelude;
 
 pub use crate::model::components::{
-    gxl_env::{GxlEnv, GxlEnvHold},
-    gxl_flow::{FlowHold, GxlFlow},
+    gxl_env::env::{GxlEnv, GxlEnvHold},
+    gxl_flow::{flow::FlowHold, flow::GxlFlow},
     gxl_mod::{GxlMod, ModHold},
     gxl_var::{RgVars, VarsHold},
 };
