+++ src/model/execution/hold.rs
@@ -2,7 +2,7 @@ use async_trait::async_trait;
 use derive_more::From;
 
 use crate::ability::GxRead;
-use crate::components::gxl_intercept::FlowRunner;
+use crate::components::gxl_flow::runner::FlowRunner;
 use crate::components::gxl_mod::ModRunner;
 use crate::components::{GxlEnv, GxlFlow, GxlMod};
 use crate::context::ExecContext;
